citizen-dashboard-backend/backend/pyproject.toml
[tool.poetry]
name = "citizen-dashboard-api"
version = "0.1.0"
description = ""
authors = ["yeralash <mishaneiman@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
Flask = "^3.0.3"
sqlalchemy = "^2.0.32"
pandas = "^2.2.2"
pyarrow = "^17.0.0"
flask-cors = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
